requirements.txt
ultralytics
numpy
matplotlib
pandas
torch
transformers
pillow
sentence_transformers
pinecone-client
pinecone
dotenv
faiss
fashion-clip
streamlit